Rework torch.hub.load() support code
